Cleaned up the algorithms a bit
