+++ test/runtests.jl
@@ -3,7 +3,7 @@ using ConstrainedSystems
 using Test
 ##using TestSetExtensions
 
-include("utils.jl")
+#include("utils.jl")
 #include("saddle.jl")
 #include("timemarching.jl")
 include("algconvergence.jl")
